24 our web application should be able to use sqlalchemy to store data (#27)
* #24 Our microservice should be able to use SQLAlchemy to access persistent data

This code make sure that sqlalchemy create read database, create tables, and create rows within a table. It will also prove that is create read the session table that is used to store sqlalchemy session information.

* #24 Our microservice should be able to use SQLAlchemy to access persiste

* placeholder for environment configuration

* #27 our web application should be able to use sqlalchemy to store data

SQLAlchemy should be used to access the local database. This will impact the applications ability to successfully connect to the myself

---------

Co-authored-by: dkbaldwin <dkbaldwin@indevelopme.com>
Co-authored-by: dkbaldwin <87735707+dkbaldwin@users.noreply.github.com>